Fix Sonar Issues - [roslyn:CS8603]
