+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/IGameObjectVisualizationContentStorage.cs
@@ -1,0 +1,12 @@
+﻿
+using Microsoft.Xna.Framework.Content;
+using Microsoft.Xna.Framework.Graphics;
+
+namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
+{
+    public interface IGameObjectVisualizationContentStorage
+    {
+        Texture2D GetConsumingEffectTexture();
+        void LoadContent(ContentManager content);
+    }
+}
