+++ Zilon.Core/CDT.LAST.MonoGameClient/ViewModels/MainScene/IPersonVisualizationContentStorage.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 
 using Microsoft.Xna.Framework.Content;
-using Microsoft.Xna.Framework.Graphics;
 
 namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
 {
@@ -16,26 +15,5 @@ namespace CDT.LAST.MonoGameClient.ViewModels.MainScene
         IEnumerable<HeadPart> GetHeadParts(string sid);
         IEnumerable<BodyPart> GetHumanParts();
         void LoadContent(ContentManager content);
-    }
-
-    public interface IGameObjectVisualizationContentStorage
-    {
-        Texture2D GetConsumingEffectTexture();
-        void LoadContent(ContentManager content);
-    }
-
-    public sealed class GameObjectVisualizationContentStorage : IGameObjectVisualizationContentStorage
-    {
-        private Texture2D? _cunsumingEffectTexture;
-
-        public Texture2D GetConsumingEffectTexture()
-        {
-            return _cunsumingEffectTexture;
-        }
-
-        public void LoadContent(ContentManager content)
-        {
-            _cunsumingEffectTexture = content.Load<Texture2D>("Sprites/effects/ConsumingEffects");
-        }
     }
 }
